Bug fixes & readability imrpovements
The script now marks champs as "invalid" if the API returns an error code of 500 in its response when attempting to hover/lock the champ. This fixes a bug where the script would think that the character is still a valid pick after having been banned. Also made the http error code handling more graceful, and some miscellaneous refactoring.