Remove program termination in main on connection loss
Will now attempt to re-parse lockfile if connection is lost. If the lockfile is not present, the client was closed, so the script really should terminate (and it will, inside connection.parse_lockfile). If the lockfile IS present, then the client is still open, and the connection was just momentarily lost.