Add lobby-creation feature & rune-sending
Added functions to create a lobby (currently supports draft, ranked solo/duo, and ranked flex), and the appropriate JS/HTML buttons and functions to access them. Also made it so that runes are sent to the client right away when the user clicks the rune checkbox on during champselect. Finally, implemented function for user-facing chamion name formatting.