Bug fix
Fixed a bug causing the script to not start. api.js was accidentally returning response['status'] (which is always undefined) instead of response['success']. Also changed get_role() in webapp.py to return 'success' = true even if the user is on the main menu and doesn't have a role. In that case, return an empty string for the role instead of returning 'success' = false